fix: functions.c - Return of function not int
The return of the function was an int and the returned value wasn't, a cast was needed.

+++ logic/functions/functions.c
@@ -27,7 +27,7 @@ int comparePrices(const void * a, const void * b) {
 	double aValue = *((double *) a);
 	double bValue = *((double *) b);
 
-	return round(aValue - bValue);
+	return (int)round(aValue - bValue);
 }
 
 void planTrip(List *l){
